requirements.txt
Flask>=3.0,<4.0
gunicorn>=22.0,<23.0
python-dotenv>=1.0,<2.0
